+++ RayCaster/RayCasterFixed.h
@@ -4,7 +4,7 @@ class RayCasterFixed : public RayCaster
 {
 public:
     void Start(uint32_t playerX, uint32_t playerY, int32_t playerA);
-    void Trace(uint32_t screenX, uint8_t* screenY, uint8_t* textureNo, uint8_t* textureX, uint32_t* textureY, uint32_t* textureStep);
+    void Trace(uint32_t screenX, uint32_t* screenY, uint8_t* textureNo, uint8_t* textureX, uint32_t* textureY, uint32_t* textureStep);
 
     RayCasterFixed();
     ~RayCasterFixed();
