+++ RayCaster/RayCasterFloat.h
@@ -6,7 +6,7 @@ class RayCasterFloat : public RayCaster
 {
 public:
     void Start(uint32_t playerX, uint32_t playerY, int32_t playerA);
-    void Trace(uint32_t screenX, uint8_t* screenY, uint8_t* textureNo, uint8_t* textureX, uint32_t* textureY, uint32_t* textureStep);
+    void Trace(uint32_t screenX, uint32_t* screenY, uint8_t* textureNo, uint8_t* textureX, uint32_t* textureY, uint32_t* textureStep);
 
     RayCasterFloat();
     ~RayCasterFloat();
